refactor(popover): file names
