+++ ui-admin/upload/src/component.tsx
